refactor(Controller): App 클래스의 기능을 Controller 와 분할 - App 클래스에서 전체적인 동작 흐름 제어 - Controller 에서 실질적으로 UI 와 도메인 로직을 연결하도록 리팩토링
